Fixed svg(heart), move categoriesSlise in folder slice to folder slices, add try-catch to categoriesSlice
