+++ src/assets/iconComponents/SvgHeart.jsx
@@ -11,13 +11,13 @@ export const SvgHeart = () => {
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
     >
-      <g clip-path="url(#clip0_6634_8345)">
+      <g clipPath="url(#clip0_6634_8345)">
         <path
           d="M39.4 32.2222C42.678 29.14 46 25.4456 46 20.6111C46 17.5317 44.7252 14.5783 42.456 12.4008C40.1868 10.2233 37.1091 9 33.9 9C30.028 9 27.3 10.0556 24 13.2222C20.7 10.0556 17.972 9 14.1 9C10.8909 9 7.8132 10.2233 5.54401 12.4008C3.27482 14.5783 2 17.5317 2 20.6111C2 25.4667 5.3 29.1611 8.6 32.2222L24 47L39.4 32.2222Z"
           stroke="#282828"
-          stroke-width="2.2"
-          stroke-linecap="round"
-          stroke-linejoin="round"
+          strokeWidth="2.2"
+          strokeLinecap="round"
+          strokeLinejoin="round"
         />
       </g>
       <defs>
