+++ src/store/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
 import productsSlice from './slices/productsSlice';
-import categoriesSlise from './slice/categoriesSlice';
+import categoriesSlise from './slices/categoriesSlice';
 
 export const store = configureStore({
   reducer: {
